[simulator][machine] Split into several files
- machine is the pojo and factory for machine
- simulator is the real machine simulator
